Custom hook
Custom hook created for code readability and reuseable
