+++ index.html
@@ -3,8 +3,9 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <link rel="icon" type="image/x-icon"
+    href="https://files.yappe.in/place/full/food-villa-family-restaurant-hotel-4642683.webp">
   <title>Foodvilla</title>
 </head>
