+++ src/component/Header.css
@@ -10,6 +10,8 @@
   padding: 15px;
   font-size: 16px;
   border-radius: 6px;
+  font-weight: 600;
+  font-size: larger;
 }
 
 .nav-items li a {
