+++ src/component/ResturantMenu.jsx
@@ -2,24 +2,12 @@ import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import Shimmer from "./Shimmer";
 import "./ResturantMenu.css";
+import useResturantMenu from "../utils/useResturantMenu";
 
 const ResturantMenu = () => {
-  const [resInfo, setResInfo] = useState(null);
   const { id } = useParams();
 
-  useEffect(() => {
-    fetchMenu();
-  }, []);
-
-  const fetchMenu = async () => {
-    const data = await fetch(
-      "https://685265a50594059b23cd54b2.mockapi.io/FoodVilla"
-    );
-    const json = await data.json();
-
-    const restaurant = json.find((res) => res.id === id);
-    setResInfo(restaurant);
-  };
+  const resInfo = useResturantMenu(id);
 
   return resInfo === null ? (
     <Shimmer />
